+++ scripts/core/helpers/fetch-all.tsx
