+++ scripts/core/ui/components/generic-form/input-types/content-filter-single-value.tsx
@@ -1,5 +1,5 @@
 import {getSelectSingleValue} from './select_single_value';
-import {fetchAll} from 'core/helpers/fetch-all-pages';
+import {fetchAll} from 'core/helpers/fetch-all';
 
 export const ContentFilterSingleValue = getSelectSingleValue(
     () => fetchAll('content_filters', {field: 'name', direction: 'ascending'})
